requirements.txt
pandas==2.0.3
plotly==5.18.0
flask==3.0.0
requests==2.31.0
python-dotenv==1.0.0
dash==2.14.2
dash-bootstrap-components==1.5.0